+++ CHANGES.rst
@@ -227,3 +227,4 @@ v1.2.9
 
 v1.2.10[unreleased]
 -----------
+* Add causes and skills to project serializers